pyproject.toml
[tool.poetry]
name = "yepcord-server"
version = "1.0.0b4"
description = "YEPCord - Free open source selfhostable fully discord-compatible chat"
authors = ["RuslanUC <dev_ruslan_uc@protonmail.com>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "yepcord"}]
repository = "https://github.com/yepcord/server"
homepage = "https://github.com/yepcord/server"
keywords = ["yepcord", "chat", "discord", "chat-server", "chat-backend", "discord-backend"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AsyncIO",
    "Framework :: Flask",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 2",
    "Framework :: Pytest",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Communications :: Chat",
    "Topic :: Internet",
    "Typing :: Typed",
]

[tool.poetry.scripts]
yepcord = "yepcord.cli:main"

[tool.poetry.dependencies]
python = "^3.9"
quart = "0.19.4"
aiofiles = "23.2.1"
websockets = "12.0"
uvicorn = "0.27.1"
aiohttp = "3.9.3"
python-magic = "0.4.27"
pillow = "10.2.0"
protobuf = "4.25.3"
python-dateutil = "2.8.2"
cryptography = "42.0.4"
emoji = "2.10.1"
six = "1.16.0"
bcrypt = "4.1.2"
quart-schema = "0.19.1"
pydantic = "2.6.1"
werkzeug = "3.0.1"
aioftp = "0.22.3"
orjson = "3.9.14"
mailers = {version = "3.0.1", extras = ["smtp"]}
propan = {version = "0.1.5.26", extras = ["async-rabbit", "async-redis"]}  # TODO: replace with faststream
redis = ">=4.6.0"
click = "8.1.7"
maxminddb = "2.5.2"
wget = "3.2"
tortoise-orm = {extras = ["aiosqlite", "asyncmy"], version = "^0.20.0"}
uvloop = "0.19.0"
async-timeout = "^4.0.3"
aerich = "^0.7.2"
yc-protobuf3-to-dict = "^0.3.0"
s3lite = "^0.1.3"
fast-depends = ">=2.1.1,<2.2.0"

[tool.poetry.group.dev.dependencies]
pytest-cov = "4.1.0"
pytest-asyncio = "0.23.5"
pyftpdlib = "1.5.8"
fake-s3 = "1.0.2"
types-protobuf = "^4.24.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
